requirements.txt
# requirements.txt

Flask==2.3.3
opencv-python-headless==4.8.1.78
numpy==1.26.2
Pillow==10.1.0
requests==2.31.0
Werkzeug==3.0.1
python-dotenv==1.0.0